Don't apply httppolicyset specific to port 80
As its ensured that the VS is listening on port 80 anyway, we don't need to only apply the rule to port 80. The advantage is following: When HTTP-to-HTTPS redirect is enabled in the ApplicationProfile/parent VirtualHosting, the rule for port 80 won't apply and token validation fails. Without this, it also works on HTTPS (443) and token verification succeeds (LetsEncrypt follows redirect from :80 to :443 for token validation)