Implemented support for VirtualHosting scenarios
This was quite tricky, as child-VH VSs use vh_domain_name and cannot be identified via fqdn and don't have services object in the result. Therefore, when child VS detected, we do a slightly different search for VSs and set the services object from the parent VS. This way the later logic is mainly untouched.